Move tracker to UI
